chore: move Onboarding to its own stories file
