+++ packages/client/src/v2-events/features/events/actions/correct/request/CorrectionRequest.stories.tsx
@@ -40,26 +40,6 @@ const tRPCMsw = createTRPCMsw<AppRouter>({
   transformer: { input: superjson, output: superjson }
 })
 
-export const Onboarding: Story = {
-  parameters: {
-    reactRouter: {
-      router: router,
-      initialPath: ROUTES.V2.EVENTS.REQUEST_CORRECTION.ONBOARDING.buildPath({
-        eventId: tennisClueMembershipEventDocument.id,
-        pageId: 'corrector'
-      })
-    },
-    msw: {
-      handlers: {
-        event: [
-          tRPCMsw.event.get.query(() => {
-            return tennisClueMembershipEventDocument
-          })
-        ]
-      }
-    }
-  }
-}
 function FormClear() {
   const form = useEventFormData()
   const requestDetails = useCorrectionRequestData()
